:bicyclist: perf: Speed-up AUR packages install step.
Trizen pulls changes by default and there does not seem to be a way
around that. Fixing https://github.com/trizen/trizen/issues/238 would
help, but that would not be enough. Since updates are handled earlier in
the script, there is no need to check for updates a second time here.
Hence, here we only pass to trizen the list of packages that are not
installed (`pacman -Qk` tests the opposite). Most of the time this list
is empty, efficiently bypassing all checks.

Fixes #21.