:adhesive_bandage: fix(bin/github.mark_as_read): Make it scale.
Before this change having a lot of PRs stalled the thing because there
was no waiting or caching implemented. Here we use caching. But it has a
flaw: we test given filters sequentially, if cache is invalidated for
one filter, all preceding filters should be rerun on the more recent
output. We cannot just take the min validity since then we would prevent
early branch out based on cached document.